# fix(sql): use xxhash hash algorithm to improve join performance