--- a/core/src/main/java/io/questdb/cairo/vm/AbstractMemoryCR.java
+++ b/core/src/main/java/io/questdb/cairo/vm/AbstractMemoryCR.java
@@ -112,8 +112,4 @@ public abstract class AbstractMemoryCR implements MemoryCR {
     public FilesFacade getFilesFacade() {
         return ff;
     }
-
-    public long getGrownLength() {
-        return grownLength;
-    }
 }