Current hash function used for `FastMap` and `CompactMap` hash table implementations has issues with quality. For instance, in case of "1".."N" string keys there is a lot of hash code collisions leading to large clusters created by the hash table due to linear probing.

This change swaps the old hash function with 64-bit variant of xxHash (http://cyan4973.github.io/xxHash).

It also gets rid of litter produced by method references in the `MemoryR#hash0()` method.

### Benchmarks

Before:

FastMapTest.testCollisionPerformance - 3 sec
CompactMapTest.testCollisionPerformanceStringKeys - 16.5 sec

```
Benchmark                                     Mode  Cnt  Score   Error  Units
QMapReadSequentialKeyBenchmark.testDirectMap  avgt    3  8.625 ± 0.536  us/op
QMapReadSequentialKeyBenchmark.testQMap       avgt    3  1.101 ± 0.058  us/op
QMapReadRandomKeyBenchmark.testDirectMap      avgt    3  0.700 ± 0.596  us/op
QMapReadRandomKeyBenchmark.testQMap           avgt    3  0.700 ± 0.170  us/op
```

After:

FastMapTest.testCollisionPerformance - 3.1 sec
CompactMapTest.testCollisionPerformanceStringKeys - 4.7 sec

```
Benchmark                                     Mode  Cnt  Score   Error  Units
QMapReadSequentialKeyBenchmark.testDirectMap  avgt    3  0.234 ± 0.067  us/op
QMapReadSequentialKeyBenchmark.testQMap       avgt    3  0.291 ± 0.095  us/op
QMapReadRandomKeyBenchmark.testDirectMap      avgt    3  0.755 ± 0.584  us/op
QMapReadRandomKeyBenchmark.testQMap           avgt    3  0.747 ± 0.461  us/op
```