@ideoma 2022-05-31
#### [PR Coverage check]

:heart_eyes: **pass** : 91 / 91 (100.00%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/Hash.java|78|78|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/FastMap.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/CompactMap.java|10|10|100.00%|